elem: drop item before taking one
This situation shouldn't occure in theory... whatever

+++ item/FallingBlock.java
@@ -36,7 +36,7 @@ public class FallingBlock extends Item{
 	@Override
 	public void update(){
 		World W = World.getInstance();
-		int i = (int)b.z;
+		int i = b.z;
 		while (i>0 && !W.hasSolidFloor(b.x, b.y, i))
 			--i;
 		int mat = W.getMaterialID(b.x, b.y, i);
@@ -46,7 +46,7 @@ public class FallingBlock extends Item{
 			W.setForm(b.x, b.y, i, f);
 			//spawn drop from mat, form
 		} else if (mat == m) {
-			W.setForm(b.x, b.y, i, (int)Math.min(form, f));
+			W.setForm(b.x, b.y, i, Math.min(form, f));
 			//spawn drop from m,max(f,form)
 		} else {
 			//spawn drop from m,f
